Move `ty::{Predicate, PredicateKind}` to their own little module (cute)
